telegram_bot/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
python-telegram-bot>=21.0
APScheduler>=3.10.0
feedparser>=6.0.0
anthropic>=0.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
